Enum embedded_hal_nb::serial::ErrorKind

#[non_exhaustive]
pub enum ErrorKind {
    Overrun,
    FrameFormat,
    Parity,
    Noise,
    Other,
}

Serial error kind.

This represents a common set of serial operation errors. HAL implementations are free to define more specific or additional error types. However, by providing a mapping to these common serial errors, generic code can still react to them.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Overrun

The peripheral receive buffer was overrun.

FrameFormat

Received data does not conform to the peripheral configuration. Can be caused by a misconfigured device on either end of the serial line.

Parity

Parity check failed.

Noise

Serial line is too noisy to read valid data.

Other

A different error occurred. The original error may contain more information.